requirements.txt
Flask>=2.3.3
Flask-Login
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
Flask-Mail
Flask-CORS>=4.0.0
PyJWT>=2.8.0
python-dotenv # Added for .env file support
aiosqlite
pandas
openpyxl
openai
requests
chromadb>=0.4.0
sentence-transformers
rank-bm25
nltk